refactor: delegate auth to core module
